requirements.txt
numpy==1.18.1
opencv-python==4.0.0.21
PyYAML==5.3
scikit-image==0.16.2
scipy==1.4.1
dlib==19.19.0
tqdm==4.43.0
imutils==0.5.3
